Refactor of propTypes
