+++ react/locales/es-AR.json
@@ -9,13 +9,16 @@
   "editor.productKitList.dot": "Punto",
   "editor.productKitList.plusIcon": "Icono Más",
   "editor.productKitList.equalsIcon": "Icono Igual",
+  "editor.productKitList.swapIcon": "Icono de Intercambio",
+  "editor.productKitList.removalIcon": "Icono de Eliminación",
   "editor.productKitList.description": "El component de kit de productos muestra una lista horizontal de productos que pertenecen al kit",
   "editor.productKitList.showListPrice": "Mostrar lista de precios",
   "editor.productKitList.showLabels": "Mostrar las etiquetas",
   "editor.productKitList.showInstallments": "Mostrar parcelas",
   "editor.productKitList.showBadge": "Mostrar badge",
   "editor.productKitList.badgeText": "Texto de badge",
-  "productKitList.numberOfProductsTitle": "Compre los {numberOfItems} productos",
+  "productKitList.takeAll": "Ileve los {numberOfItems} productos",
   "productKitList.mountYourKit": "Monte su kit",
-  "productKitList.swapItem": "Intercambiar"
+  "productKitList.swapItem": "Intercambiar",
+  "productKitList.buyKit": "Comprar Kit"
 }
