+++ react/locales/pt-BR.json
@@ -9,13 +9,16 @@
   "editor.productKitList.showDots": "Mostrar Pontos",
   "editor.productKitList.dot": "Ponto",
   "editor.productKitList.plusIcon": "Ícone de Mais",
-  "editor.productKitList.equalsIcon": "Íconde de Igual",
+  "editor.productKitList.equalsIcon": "Ícone de Igual",
+  "editor.productKitList.swapIcon": "Ícone de Troca",
+  "editor.productKitList.removalIcon": "Ícone de Remoção",
   "editor.productKitList.showListPrice": "Mostrar lista de preços",
   "editor.productKitList.showLabels": "Mostrar rótulos",
   "editor.productKitList.showInstallments": "Mostrar parcelamentos",
   "editor.productKitList.showBadge": "Mostrar badge",
   "editor.productKitList.badgeText": "Texto da badge",
-  "productKitList.numberOfProductsTitle": "Compre os {numberOfItems} produtos",
+  "productKitList.takeAll": "Leve os {numberOfItems} produtos",
   "productKitList.mountYourKit": "Monte seu kit",
-  "productKitList.swapItem": "Trocar"
+  "productKitList.swapItem": "Trocar",
+  "productKitList.buyKit": "Comprar Kit"
 }
