+++ react/props/productKitDefaultProps.js
@@ -1,0 +1,29 @@
+import defaultPlusIcon from '../images/default-plus-icon.svg'
+import defaultEqualsIcon from '../images/default-equals-icon.svg'
+import defaultSwapIcon from '../images/default-swap-icon.svg'
+import defaultRemovalIcon from '../images/default-removal-icon.svg'
+
+export default {
+  /** Slick Slider */
+  showArrows: true,
+  showDots: true,
+
+  /** Product Summary */
+  showListPrice: true,
+  showLabel: false,
+  showInstallments: false,
+  showBadge: true,
+  showCollections: true,
+
+  /** Component Flag Operations */
+  allowSwap: true,
+  allowRemoval: true,
+
+  /** Operation Icons */
+  swapIcon: defaultSwapIcon,
+  removalIcon: defaultRemovalIcon,
+
+  /** Separator Icons */
+  plusIcon: defaultPlusIcon,
+  equalsIcon: defaultEqualsIcon,
+}
